marta_transit_dashboard/requirements.txt
Flask
requests
openmeteo-requests
requests-cache
retry-requests
pandas
pillow
protobuf
gtfs-realtime-bindings 